% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_operations.R
\name{delete_organization}
\alias{delete_organization}
\title{Deletes the organization}
\usage{
delete_organization()
}
\description{
Deletes the organization. You can delete an organization only by using credentials from the master account. The organization must be empty of member accounts.
}
