% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_operations.R
\name{describe_organization}
\alias{describe_organization}
\title{Retrieves information about the organization that the user's account belongs to}
\usage{
describe_organization()
}
\description{
Retrieves information about the organization that the user's account belongs to.
}
\details{
This operation can be called from any account in the organization.

Even if a policy type is shown as available in the organization, it can be disabled separately at the root level with DisablePolicyType. Use ListRoots to see the status of policy types for a specified root.
}
\examples{
# The following example shows how to request information about the current
# user's organization:/n/n
describe_organization()

}
