% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_operations.R
\name{describe_organizational_unit}
\alias{describe_organizational_unit}
\title{Retrieves information about an organizational unit (OU)}
\usage{
describe_organizational_unit(OrganizationalUnitId)
}
\arguments{
\item{OrganizationalUnitId}{[required] The unique identifier (ID) of the organizational unit that you want details about. You can get the ID from the ListOrganizationalUnitsForParent operation.

The \href{http://wikipedia.org/wiki/regex}{regex pattern} for an organizational unit ID string requires "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that contains the OU) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.}
}
\description{
Retrieves information about an organizational unit (OU).
}
\details{
This operation can be called only from the organization's master account.
}
\section{Accepted Parameters}{
\preformatted{describe_organizational_unit(
  OrganizationalUnitId = "string"
)
}
}

\examples{
# The following example shows how to request details about an OU:/n/n
\donttest{describe_organizational_unit(
  OrganizationalUnitId = "ou-examplerootid111-exampleouid111"
)}

}
