% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_operations.R
\name{describe_policy}
\alias{describe_policy}
\title{Retrieves information about a policy}
\usage{
describe_policy(
  PolicyId = "string"
)
}
\arguments{
\item{PolicyId}{[required] The unique identifier (ID) of the policy that you want details about. You can get the ID from the ListPolicies or ListPoliciesForTarget operations.

The \href{http://wikipedia.org/wiki/regex}{regex pattern} for a policy ID string requires "p-" followed by from 8 to 128 lower-case letters or digits.}
}
\description{
Retrieves information about a policy.
}
\details{
This operation can be called only from the organization's master account.
}
\examples{
# The following example shows how to request information about a
# policy:/n/n
describe_policy(
  PolicyId = "p-examplepolicyid111"
)

}
