% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_operations.R
\name{detach_policy}
\alias{detach_policy}
\title{Detaches a policy from a target root, organizational unit (OU), or account}
\usage{
detach_policy(
  PolicyId = "string",
  TargetId = "string"
)
}
\arguments{
\item{PolicyId}{[required] The unique identifier (ID) of the policy you want to detach. You can get the ID from the ListPolicies or ListPoliciesForTarget operations.

The \href{http://wikipedia.org/wiki/regex}{regex pattern} for a policy ID string requires "p-" followed by from 8 to 128 lower-case letters or digits.}

\item{TargetId}{[required] The unique identifier (ID) of the root, OU, or account from which you want to detach the policy. You can get the ID from the ListRoots, ListOrganizationalUnitsForParent, or ListAccounts operations.

The \href{http://wikipedia.org/wiki/regex}{regex pattern} for a target ID string requires one of the following:
\itemize{
\item Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits.
\item Account: a string that consists of exactly 12 digits.
\item Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.
}}
}
\description{
Detaches a policy from a target root, organizational unit (OU), or account. If the policy being detached is a service control policy (SCP), the changes to permissions for IAM users and roles in affected accounts are immediate.
}
\details{
\strong{Note:} Every root, OU, and account must have at least one SCP attached. If you want to replace the default \code{FullAWSAccess} policy with one that limits the permissions that can be delegated, then you must attach the replacement policy before you can remove the default one. This is the authorization strategy of \href{http://docs.aws.amazon.com/organizations/latest/userguide/orgs_manage_policies_about-scps.html#orgs_policies_whitelist}{whitelisting}. If you instead attach a second SCP and leave the \code{FullAWSAccess} SCP still attached, and specify \code{"Effect": "Deny"} in the second SCP to override the \code{"Effect": "Allow"} in the \code{FullAWSAccess} policy (or any other attached SCP), then you are using the authorization strategy of \href{http://docs.aws.amazon.com/organizations/latest/userguide/orgs_manage_policies_about-scps.html#orgs_policies_blacklist}{blacklisting}.

This operation can be called only from the organization's master account.
}
\examples{
# The following example shows how to detach a policy from an OU:/n/n
detach_policy(
  PolicyId = "p-examplepolicyid111",
  TargetId = "ou-examplerootid111-exampleouid111"
)

}
