% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_operations.R
\name{invite_account_to_organization}
\alias{invite_account_to_organization}
\title{Sends an invitation to another account to join your organization as a member account}
\usage{
invite_account_to_organization(Target, Notes = NULL)
}
\arguments{
\item{Notes}{Additional information that you want to include in the generated email to the recipient account owner.}
}
\description{
Sends an invitation to another account to join your organization as a member account. Organizations sends email on your behalf to the email address that is associated with the other account's owner. The invitation is implemented as a Handshake whose details are in the response.
}
\details{
\itemize{
\item You can invite AWS accounts only from the same seller as the master account. For example, if your organization's master account was created by Amazon Internet Services Pvt. Ltd (AISPL), an AWS seller in India, then you can only invite other AISPL accounts to your organization. You can't combine accounts from AISPL and AWS, or any other AWS seller. For more information, see \href{http://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/useconsolidatedbilliing-India.html}{Consolidated Billing in India}.
\item If you receive an exception that indicates that you exceeded your account limits for the organization or that the operation failed because your organization is still initializing, wait one hour and then try again. If the error persists after an hour, then contact \href{https://console.aws.amazon.com/support/home#/}{AWS Customer Support}.
}

This operation can be called only from the organization's master account.
}
\section{Accepted Parameters}{
\preformatted{invite_account_to_organization(
  Target = list(
    Id = "string",
    Type = "ACCOUNT"|"ORGANIZATION"|"EMAIL"
  ),
  Notes = "string"
)
}
}

\examples{
# The following example shows the admin of the master account owned by
# bill@example.com inviting the account owned by juan@example.com to join
# an organization.
\donttest{invite_account_to_organization(
  Notes = "This is a request for Juan's account to join Bill's organization",
  Target = list(
    Id = "juan@example.com",
    Type = "EMAIL"
  )
)}

}
