% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_operations.R
\name{list_aws_service_access_for_organization}
\alias{list_aws_service_access_for_organization}
\title{Returns a list of the AWS services that you enabled to integrate with your organization}
\usage{
list_aws_service_access_for_organization(NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{NextToken}{Use this parameter if you receive a \code{NextToken} response in a previous request that indicates that there is more output available. Set it to the value of the previous call's \code{NextToken} response to indicate where the output should continue from.}

\item{MaxResults}{(Optional) Use this to limit the number of results you want included per page in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the \code{NextToken} response element is present and has a value (is not null). Include that value as the \code{NextToken} request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check \code{NextToken} after every operation to ensure that you receive all of the results.}
}
\description{
Returns a list of the AWS services that you enabled to integrate with your organization. After a service on this list creates the resources that it requires for the integration, it can perform operations on your organization and its accounts.
}
\details{
For more information about integrating other services with AWS Organizations, including the list of services that currently work with Organizations, see \href{http://docs.aws.amazon.com/organizations/latest/userguide/orgs_integrate_services.html}{Integrating AWS Organizations with Other AWS Services} in the \emph{AWS Organizations User Guide}.

This operation can be called only from the organization's master account.
}
\examples{

}
