% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_operations.R
\name{list_create_account_status}
\alias{list_create_account_status}
\title{Lists the account creation requests that match the specified status that is currently being tracked for the organization}
\usage{
list_create_account_status(
  States = list(
    "IN_PROGRESS"|"SUCCEEDED"|"FAILED"
  ),
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{States}{A list of one or more states that you want included in the response. If this parameter is not present, then all requests are included in the response.}

\item{NextToken}{Use this parameter if you receive a \code{NextToken} response in a previous request that indicates that there is more output available. Set it to the value of the previous call's \code{NextToken} response to indicate where the output should continue from.}

\item{MaxResults}{(Optional) Use this to limit the number of results you want included per page in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the \code{NextToken} response element is present and has a value (is not null). Include that value as the \code{NextToken} request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check \code{NextToken} after every operation to ensure that you receive all of the results.}
}
\description{
Lists the account creation requests that match the specified status that is currently being tracked for the organization.
}
\details{
Always check the \code{NextToken} response parameter for a \code{null} value when calling a \code{List*} operation. These operations can occasionally return an empty set of results even when there are more results available. The \code{NextToken} response parameter value is \code{null} \emph{only} when there are no more results to display.

This operation can be called only from the organization's master account.
}
\examples{
# The following example shows a user requesting a list of only the
# completed account creation requests made for the current organization:
list_create_account_status(
  States = list(
    "SUCCEEDED"
  )
)

# The following example shows a user requesting a list of only the
# in-progress account creation requests made for the current organization:
list_create_account_status(
  States = list(
    "IN_PROGRESS"
  )
)

}
