% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_operations.R
\name{list_handshakes_for_account}
\alias{list_handshakes_for_account}
\title{Lists the current handshakes that are associated with the account of the requesting user}
\usage{
list_handshakes_for_account(Filter = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{Filter}{Filters the handshakes that you want included in the response. The default is all types. Use the \code{ActionType} element to limit the output to only a specified type, such as \code{INVITE}, \code{ENABLE-FULL-CONTROL}, or \code{APPROVE-FULL-CONTROL}. Alternatively, for the \code{ENABLE-FULL-CONTROL} handshake that generates a separate child handshake for each member account, you can specify \code{ParentHandshakeId} to see only the handshakes that were generated by that parent request.}

\item{NextToken}{Use this parameter if you receive a \code{NextToken} response in a previous request that indicates that there is more output available. Set it to the value of the previous call's \code{NextToken} response to indicate where the output should continue from.}

\item{MaxResults}{(Optional) Use this to limit the number of results you want included per page in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the \code{NextToken} response element is present and has a value (is not null). Include that value as the \code{NextToken} request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check \code{NextToken} after every operation to ensure that you receive all of the results.}
}
\description{
Lists the current handshakes that are associated with the account of the requesting user.
}
\details{
Handshakes that are ACCEPTED, DECLINED, or CANCELED appear in the results of this API for only 30 days after changing to that state. After that they are deleted and no longer accessible.

Always check the \code{NextToken} response parameter for a \code{null} value when calling a \code{List*} operation. These operations can occasionally return an empty set of results even when there are more results available. The \code{NextToken} response parameter value is \code{null} \emph{only} when there are no more results to display.

This operation can be called from any account in the organization.
}
\examples{
# The following example shows you how to get a list of handshakes that are
# associated with the account of the credentials used to call the
# operation:
list_handshakes_for_account()

}
