% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_operations.R
\name{list_policies_for_target}
\alias{list_policies_for_target}
\title{Lists the policies that are directly attached to the specified target root, organizational unit (OU), or account}
\usage{
list_policies_for_target(TargetId, Filter, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{TargetId}{[required] The unique identifier (ID) of the root, organizational unit, or account whose policies you want to list.

The \href{http://wikipedia.org/wiki/regex}{regex pattern} for a target ID string requires one of the following:
\itemize{
\item Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits.
\item Account: a string that consists of exactly 12 digits.
\item Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.
}}

\item{Filter}{[required] The type of policy that you want to include in the returned list.}

\item{NextToken}{Use this parameter if you receive a \code{NextToken} response in a previous request that indicates that there is more output available. Set it to the value of the previous call's \code{NextToken} response to indicate where the output should continue from.}

\item{MaxResults}{(Optional) Use this to limit the number of results you want included per page in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the \code{NextToken} response element is present and has a value (is not null). Include that value as the \code{NextToken} request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check \code{NextToken} after every operation to ensure that you receive all of the results.}
}
\description{
Lists the policies that are directly attached to the specified target root, organizational unit (OU), or account. You must specify the policy type that you want included in the returned list.
}
\details{
Always check the \code{NextToken} response parameter for a \code{null} value when calling a \code{List*} operation. These operations can occasionally return an empty set of results even when there are more results available. The \code{NextToken} response parameter value is \code{null} \emph{only} when there are no more results to display.

This operation can be called only from the organization's master account.
}
\section{Accepted Parameters}{
\preformatted{list_policies_for_target(
  TargetId = "string",
  Filter = "SERVICE_CONTROL_POLICY",
  NextToken = "string",
  MaxResults = 123
)
}
}

\examples{
# The following example shows how to get a list of all service control
# policies (SCPs) of the type specified by the Filter parameter, that are
# directly attached to an account. The returned list does not include
# policies that apply to the account because of inheritance from its
# location in an OU hierarchy:/n/n
\donttest{list_policies_for_target(
  Filter = "SERVICE_CONTROL_POLICY",
  TargetId = "444444444444"
)}

}
