% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.organizations_package.R
\docType{package}
\name{paws.organizations-package}
\alias{paws.organizations}
\alias{paws.organizations-package}
\title{paws.organizations: AWS Organizations}
\description{
AWS Organizations is a web service that enables you
    to consolidate your multiple AWS accounts into an *organization* and
    centrally manage your accounts and their resources.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
